requirements.txt
python-docx==1.1.0
PyPDF2==3.0.1
readability-lxml>=0.8.1
beautifulsoup4==4.12.2
requests==2.31.0
lxml>=4.9.0
python-magic>=0.4.27
tqdm>=4.65.0
fastapi>=0.115.2
uvicorn>=0.24.0
jinja2==3.1.2
python-multipart>=0.0.18
aiofiles
python-dotenv==1.0.0
# MiniMax TTS 相关依赖
websockets>=11.0.3
werkzeug
# 内容提取相关依赖已在前面列出
openai==1.6.1
langchain==0.0.350
langchain-openai==0.0.2
pydub==0.25.1
'volcengine-python-sdk[ark]' 